Fix with second window
